Move HFSS tests to own module
